Implemented the wait_util CTRL and refactored base ARI between lit/ref parts. Clarified TP value datums.
